added login

+++ TeachersGuide/Controllers/ClientController.cs
@@ -15,7 +15,7 @@ namespace TeachersGuide.Contrillers
         private IInterventionsModifiedRepository _interventionsModifiedRepository;      //NOT READONLY !!
         private IBehaviorPageTowRepository _behaviorPageTowRepository;
         private IBehaviourPageOneRepository _behaviourPageOneRepository;
-        
+        private static bool authenticationToken = false;
         public ClientController(IFeedBackRepository feedBackRepository,
                                 IInterventionsModifiedRepository interventionsModifiedRepository,
                                 IBehaviorPageTowRepository behaviorPageTowRepository,
@@ -33,93 +33,148 @@ namespace TeachersGuide.Contrillers
         }
         public IActionResult Page()
         {
-            ViewData["stats"] = _behaviorPageTowRepository.GetAllBehaviorPageTows();
-            IEnumerable<FeedBack> _feedBack = _feedBackRepository.GetAll();
-            var clientViewModel = new ClientViewModel()
+            if (TempData["sendFlag"] != null) { 
+                authenticationToken = (bool)TempData["sendFlag"];
+            }
+            if (authenticationToken) { 
+                ViewData["stats"] = _behaviorPageTowRepository.GetAllBehaviorPageTows();
+                IEnumerable<FeedBack> _feedBack = _feedBackRepository.GetAll();
+                var clientViewModel = new ClientViewModel()
+                {
+                    feedBacks = _feedBack.ToList()
+                };
+                ViewData["feedback"] = _feedBack;
+                return View();
+            }
+            else
             {
-                feedBacks = _feedBack.ToList()
-            };
-            ViewData["feedback"] = _feedBack;
-            return View();
+                return BadRequest();
+            }
         }
         public IActionResult Feedback()
         {
-            IEnumerable<FeedBack> _feedBack = _feedBackRepository.GetAll();
-            var clientViewModel = new ClientViewModel()
-            {
+            if (authenticationToken) { 
+                IEnumerable<FeedBack> _feedBack = _feedBackRepository.GetAll();
+                var clientViewModel = new ClientViewModel()
+                {
 
-                feedBacks = _feedBack.ToList()
-            };
-            return View(clientViewModel);
+                    feedBacks = _feedBack.ToList()
+                };
+                return View(clientViewModel);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         public IActionResult Statistics()
         {
-            IEnumerable<BehaviorPageTow> _bPT = _behaviorPageTowRepository.GetAllBehaviorPageTows();
-            return View(_bPT);
+            if (authenticationToken) { 
+                IEnumerable<BehaviorPageTow> _bPT = _behaviorPageTowRepository.GetAllBehaviorPageTows();
+                return View(_bPT);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
 
         public IActionResult EditBehaviour_P1()
         {
+            if (authenticationToken) { 
             IEnumerable<BehaviourPageOne> _bPO = _behaviourPageOneRepository.GetAll();
             return View(_bPO);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
 
         [HttpGet]
         public IActionResult Edit_BPO(long id)
         {
+            if (authenticationToken) { 
             BehaviourPageOne bPO = _behaviourPageOneRepository.GetById(id);
             return View(bPO);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPost]
         public IActionResult Edit_BPO(BehaviourPageOne bPO)
         {
-            try
-            {
-                if (ModelState.IsValid)
+            if (authenticationToken) { 
+                try
                 {
-                    _behaviourPageOneRepository.edit(bPO);
-                    IEnumerable<BehaviourPageOne> _bPO = _behaviourPageOneRepository.GetAll();
-                    return View("EditBehaviour_P1", _bPO);
+                    if (ModelState.IsValid)
+                    {
+                        _behaviourPageOneRepository.edit(bPO);
+                        IEnumerable<BehaviourPageOne> _bPO = _behaviourPageOneRepository.GetAll();
+                        return View("EditBehaviour_P1", _bPO);
+                    }
+                    else
+                    {
+                        return View();
+                    }
                 }
-                else
+                catch (Exception e)
                 {
                     return View();
                 }
             }
-            catch (Exception e)
-            {
-                return View();
+            else { return BadRequest();
             }
-
         }
 
         [HttpGet]
         public IActionResult Delete_BPO(long id)
         {
-            BehaviourPageOne bPO = _behaviourPageOneRepository.GetById(id);
-            return View(bPO);
+            if (authenticationToken) { 
+                BehaviourPageOne bPO = _behaviourPageOneRepository.GetById(id);
+                return View(bPO);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPost]
         public IActionResult Delete_BPO(BehaviourPageOne bPO)
         {
+            if (authenticationToken) { 
             _behaviourPageOneRepository.delete(bPO.Id);
             IEnumerable<BehaviourPageOne> _iM = _behaviourPageOneRepository.GetAll();
             return RedirectToAction("EditBehaviour_P1");
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         public IActionResult CreateNewBPO ()
         {
-            return View();
+            if (authenticationToken) { 
+                return View();
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPost]
         public IActionResult CreateNewBPO(BehaviourPageOne behaviourPageOne)
         {
+            if (authenticationToken) { 
             try
             {
                 if (ModelState.IsValid)
@@ -138,23 +193,38 @@ namespace TeachersGuide.Contrillers
             {
                 return View();
             }
+            }
+            else { return BadRequest(); }
         }
 
         public IActionResult Details_BPO (long Id)
         {
+            if (authenticationToken) { 
             IEnumerable<BehaviorPageTow> bPT = _behaviorPageTowRepository.GetBehaviorPageTows(Id).ToList();
             return View(bPT);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
         [HttpGet]
         public IActionResult Edit_BPT(long id)
         {
+            if (authenticationToken) { 
             BehaviorPageTow bPT = _behaviorPageTowRepository.GetById(id);
             return View(bPT);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPost]
         public IActionResult Edit_BPT(BehaviorPageTow bPT)
         {
+            if (authenticationToken) { 
             try
             {
                 if (ModelState.IsValid)
@@ -172,35 +242,58 @@ namespace TeachersGuide.Contrillers
             {
                 return View();
             }
-
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpGet]
         public IActionResult Delete_BPT(long id)
         {
+            if (authenticationToken) { 
             BehaviorPageTow bPT = _behaviorPageTowRepository.GetById(id);
             return View(bPT);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPost]
         public IActionResult Delete_BPT(BehaviorPageTow bPT)
         {
+            if (authenticationToken) { 
             _behaviorPageTowRepository.delete(bPT.Id);
             IEnumerable<BehaviorPageTow> _bPT = _behaviorPageTowRepository.GetBehaviorPageTows(bPT.BPOId).ToList();
             return View("Details_BPO", _bPT);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         public IActionResult CreateNewBPT(long Id)
         {
+            if (authenticationToken) { 
             var BPT = new BehaviorPageTow();
             BPT.BPOId = Id;
             BPT.Count = 0;
             return View(BPT);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPost]
         public IActionResult CreateNewBPT(BehaviorPageTow behaviorPageTow)
         {
+            if (authenticationToken) { 
             behaviorPageTow.Id = 0;
             try
             {
@@ -219,32 +312,48 @@ namespace TeachersGuide.Contrillers
             {
                 return View();
             }
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         
             public IActionResult Details_BPT(long Id,long data)
         {
+            if (authenticationToken) { 
+
             ViewBag.rout = data;
             ViewBag.id = Id;
             IEnumerable<InterventionsModified> _iM = _interventionsModifiedRepository.GetCategory(Id).ToList();
 
             return View(_iM);
+            }
+            else { return BadRequest(); }
         }
 
         public IActionResult CreateNewIntervention(long Id, long data)
         {
+            if (authenticationToken) { 
             ViewBag.rout = data;
             ViewBag.id = Id;
             ViewBag.Flag = 1;
             InterventionsModified interventionsModified = new InterventionsModified();
             interventionsModified.BPTid = Id;
             return View("CreateNewIntervention", interventionsModified);
-            //return View("Details_BPT", interventionsModified);
+                //return View("Details_BPT", interventionsModified);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPost]
         public IActionResult CreateNewIntervention(InterventionsModified interventionsModified, long data)
         {
+            if (authenticationToken) { 
             @ViewBag.rout = data;
             ViewBag.id = interventionsModified.BPTid;
             interventionsModified.id = 0;
@@ -266,6 +375,11 @@ namespace TeachersGuide.Contrillers
             {
                 return View();
             }
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
 
@@ -275,55 +389,87 @@ namespace TeachersGuide.Contrillers
         [HttpGet]
         public IActionResult Delete(long id, long data)
         {
+            if (authenticationToken) { 
             ViewBag.rout = data;
             InterventionsModified bPT = _interventionsModifiedRepository.getId(id);
             ViewBag.id = bPT.BPTid;
             return View(bPT);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPost]
         public IActionResult Delete(InterventionsModified bPT, long id, long data)
         {
+            if (authenticationToken) { 
             ViewBag.rout = data;
             ViewBag.id = bPT.BPTid;
             _interventionsModifiedRepository.delete(bPT.id);
             InterventionsModified iM = new InterventionsModified();
             return View(iM);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
         [HttpGet]
         public IActionResult Edit(long id, long data)
         {
+            if (authenticationToken) { 
             ViewBag.rout = data;
             InterventionsModified bPT = _interventionsModifiedRepository.getId(id);
             ViewBag.id = bPT.BPTid;
             ViewBag.Flag = 1;
             return View(bPT);
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPost]
         public IActionResult Edit(InterventionsModified interventionsModified, long id, long data)
         {
-            ViewBag.rout = data;
-            ViewBag.id = interventionsModified.BPTid;
-            
-            try
+            if (authenticationToken)
             {
-                if (ModelState.IsValid)
+                ViewBag.rout = data;
+                ViewBag.id = interventionsModified.BPTid;
+
+                try
                 {
-                    _interventionsModifiedRepository.edit(interventionsModified);
-                    ViewBag.Flag = 0;
-                    return View();
+                    if (ModelState.IsValid)
+                    {
+                        _interventionsModifiedRepository.edit(interventionsModified);
+                        ViewBag.Flag = 0;
+                        return View();
+                    }
+                    else
+                    {
+                        return View();
+                    }
                 }
-                else
+                catch (Exception e)
                 {
                     return View();
                 }
             }
-            catch (Exception e)
+            else
             {
-                return View();
+                return BadRequest();
             }
 
+        }
+
+        [HttpGet]
+        public IActionResult logout()
+        {
+            authenticationToken = false;
+            return Ok();
         }
        
     }
